Remove more html generation in Java
