+++ Plan/common/src/test/java/com/djrapitops/plan/storage/database/queries/ExtensionsDatabaseTest.java
@@ -18,7 +18,6 @@ package com.djrapitops.plan.storage.database.queries;
 
 import com.djrapitops.plan.component.Component;
 import com.djrapitops.plan.component.ComponentService;
-import com.djrapitops.plan.delivery.rendering.html.structure.HtmlTable;
 import com.djrapitops.plan.extension.CallEvents;
 import com.djrapitops.plan.extension.DataExtension;
 import com.djrapitops.plan.extension.ExtensionSvc;
@@ -52,10 +51,7 @@ import utilities.RandomData;
 import utilities.TestConstants;
 import utilities.TestErrorLogger;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
+import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -200,9 +196,11 @@ public interface ExtensionsDatabaseTest extends DatabaseTestPreparer {
 
         List<ExtensionTableData> tableData = tabData.getTableData();
         assertEquals(1, tableData.size());
-        HtmlTable table = tableData.get(0).getHtmlTable();
-        String result = table.toHtml();
-        assertTrue(result.contains("<tbody><tr><td>Group</td><td>1</td></tr></tbody>"), result);
+        Table table = tableData.get(0).getTable();
+        List<Object[]> expected = new ArrayList<>();
+        expected.add(new Object[]{"Group", 1});
+        List<Object[]> result = table.getRows();
+        assertEquals(expected, result);
     }
 
     @Test
@@ -330,16 +328,14 @@ public interface ExtensionsDatabaseTest extends DatabaseTestPreparer {
         assertEquals(1, tableData.size());
         ExtensionTableData table = tableData.get(0);
 
-        HtmlTable expected = HtmlTable.fromExtensionTable(
-                Table.builder()
-                        .columnOne("first", Icon.called("gavel").build())
-                        .columnTwo("second", Icon.called("what").build())
-                        .columnThree("third", Icon.called("question").build())
-                        .addRow("value", 3, 0.5, 400L)
-                        .build(),
-                com.djrapitops.plan.delivery.rendering.html.icon.Color.AMBER);
+        Table expected = Table.builder()
+                .columnOne("first", Icon.called("gavel").build())
+                .columnTwo("second", Icon.called("what").build())
+                .columnThree("third", Icon.called("question").build())
+                .addRow("value", 3, 0.5, 400L)
+                .build();
 
-        assertEquals(expected.toHtml(), table.getHtmlTable().toHtml());
+        assertEquals(expected, table.getTable());
     }
 
     @Test
@@ -364,16 +360,14 @@ public interface ExtensionsDatabaseTest extends DatabaseTestPreparer {
         assertEquals(1, tableData.size());
         ExtensionTableData table = tableData.get(0);
 
-        HtmlTable expected = HtmlTable.fromExtensionTable(
-                Table.builder()
-                        .columnOne("first", Icon.called("gavel").build())
-                        .columnTwo("second", Icon.called("what").build())
-                        .columnThree("third", Icon.called("question").build())
-                        .addRow("value", 3, 0.5, 400L)
-                        .build(),
-                com.djrapitops.plan.delivery.rendering.html.icon.Color.AMBER);
+        Table expected = Table.builder()
+                .columnOne("first", Icon.called("gavel").build())
+                .columnTwo("second", Icon.called("what").build())
+                .columnThree("third", Icon.called("question").build())
+                .addRow("value", 3, 0.5, 400L)
+                .build();
 
-        assertEquals(expected.toHtml(), table.getHtmlTable().toHtml());
+        assertEquals(expected, table.getTable());
     }
 
     @Test
